+++ IsaacsIOSLibrary/Categories/NSDate/NSDate+Compare.m
@@ -42,7 +42,7 @@
     return [thisDate isEqualToDate:yesterday];
 }
 
-- (int)daysSince:(NSDate*)other {
+- (NSInteger)daysSince:(NSDate*)other {
     NSCalendar *gregorianCalendar = [[NSCalendar alloc] initWithCalendarIdentifier:NSGregorianCalendar];
     NSDateComponents *components = [gregorianCalendar components:NSDayCalendarUnit
                                                         fromDate:other
